+++ lib/rspec/rails/matchers/be_new_record.rb
@@ -5,6 +5,14 @@ module RSpec::Rails::Matchers
     def matches?(actual)
       !actual.persisted?
     end
+
+    def failure_message_for_should
+      "expected #{actual.inspect} to be a new record, but was persisted"
+    end
+
+    def failure_message_for_should_not
+      "expected #{actual.inspect} to be persisted, but was a new record"
+    end
   end
 
   # Passes if actual returns `false` for `persisted?`.